+++ runtime/src/ctx.c
@@ -252,7 +252,7 @@ bl_val_t* bl_ctx_set(bl_val_t* ctx, bl_val_t* key, bl_val_t* val) {
       }
    }*/
    if(key->sym_id >= ctx->vals_count) {
-     uint16_t old_count = ctx->vals_count;
+     uint64_t old_count = ctx->vals_count;
      ctx->vals_count = key->sym_id+8;
      bl_val_t** old_vals = ctx->vals;
      ctx->vals = (bl_val_t**)GC_MALLOC(sizeof(bl_val_t*)*ctx->vals_count);
